refactored model prediciton API

+++ OrchardWatch-ML/Prediction/model_manager.py
@@ -2,8 +2,8 @@ import sys
 import os
 import pickle
 import shutil
-
-
+import json
+from keras_model import *
 
 os.environ['KMP_DUPLICATE_LIB_OK']='True'
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
@@ -16,8 +16,9 @@ types_file = "model_types"
 model_dir = "/Users/dhruvilgala/Desktop/320-F19-Track-I/OrchardWatch-ML/models/"
 
 # comment after done once
-model_types = set(["keras", "pytorch"])
-pickle.dump(model_types, open(types_file, "wb"))
+if not os.path.exists(types_file):
+	model_types = set(["keras", "pytorch"])
+	pickle.dump(model_types, open(types_file, "wb"))
 
 def main():
 	# TODO: print expected arguments in case of error
@@ -25,24 +26,37 @@ def main():
 	funct_flag = sys.argv[1]
 
 	model_types = pickle.load(open(types_file, "rb"))
+
+	if os.path.exists(dict_file):
+		model_dict = pickle.load(open(dict_file, "rb"))
+	else:
+		model_dict = {}
 	
 	if funct_flag == "upload":
 		model_url = sys.argv[2]
 		model_name = sys.argv[3]
 		model_type = sys.argv[4]
 
-		if os.path.exists(dict_file):
-			model_dict = pickle.load(open(dict_file, "rb"))
+		upload_model(model_types, model_dict, model_url, model_name, model_type)
+
+	elif funct_flag == "list":
+		get_models(model_dict)
+
+	elif funct_flag == "predict":
+		model_name = sys.argv[2]
+		img_url = sys.argv[3]
+
+		if model_name not in model_dict.keys():
+			raise NameError("Model with name " + model_name + " does not exist.")
+
 		else:
-			model_dict = {}
+			model_type = model_dict[model_name]
+			if model_type == "keras":
+				predict_keras(os.path.join(model_dir, model_name), img_url)
 
-		upload_model(model_dict, model_url, model_name, model_type)
 
-	# elif funct_flag == "predict":
-
-def upload_model(model_dict, model_url, model_name, model_type):
+def upload_model(model_types, model_dict, model_url, model_name, model_type):
 	if model_type not in model_types:
-		print("Error: model type not supported. Please choose either 'keras' or 'pytorch'.")
 		raise NameError("Please select either 'keras' or 'pytorch'.")
 
 	# change to boto3 on EC2
@@ -50,8 +64,15 @@ def upload_model(model_dict, model_url, model_name, model_type):
 	
 	if model_fname:
 		print("Model uploaded successfully to " + model_fname)
-		model_dict[model_name] = (model_type, model_name)
+		model_dict[model_name] = model_type
 		pickle.dump(model_dict, open(dict_file, "wb"))
+
+def get_models(model_dict):
+	if len(model_dict) == 0:
+		raise RuntimeError("No models available, please upload a model.")
+	else:
+		model_json = json.dumps(model_dict)
+		print(model_json)
 
 
 if __name__== "__main__":
